+++ server/routes/notes.js
@@ -1,12 +1,13 @@
 const express = require("express");
 const router = express.Router();
 var CryptoJS = require("crypto-js");
+const auth = require("../middleware/auth");
 
 const Notes = require("./../models/Notes");
 
 //@route    GET /notes/:id
 //@desc     Get all the notes of userId :id
-router.get("/:id", async (req, res) => {
+router.get("/:id", auth, async (req, res) => {
   try {
     const allData = [];
     const allNotes = await Notes.find({ userId: req.params.id });
@@ -24,7 +25,7 @@ router.get("/:id", async (req, res) => {
 
 //@route    POST /notes
 //@desc     Add new note
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
   let note = {
     noteName: req.body.noteName,
     note: req.body.note,
@@ -46,7 +47,7 @@ router.post("/", async (req, res) => {
 
 //@route    PATCH /notes/:id
 //@desc     Edit NOTE
-router.patch("/:id", async (req, res) => {
+router.patch("/:id", auth, async (req, res) => {
   try {
     const note = await Notes.findByIdAndUpdate(req.params.id);
     const decryptedData = CryptoJS.AES.decrypt(note.data, "Secret key 123");
@@ -81,9 +82,12 @@ router.patch("/:id", async (req, res) => {
 
 //@route    DELETE /notes/:id
 //@desc     Delete a NOTE
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
   try {
     let note = await Notes.findById(req.params.id);
+    if (note.userId.toString() !== req.user._id.toString()) {
+      return res.status(400).json({ error: "Unauthorized access" });
+    }
     await note.deleteOne();
     res.status(201).json({ id: req.params.id });
   } catch (error) {
